tcg/plugins/dyncount: add capstone based dyncount plugin
The dyncount plugin counts dynamic instructions per
mnemonics and per groups as available in capstone
library.
The plugin support for now x86_64 and i386 guests and
relied on libcapstone >= 3.0.

Add --enable/--disable-casptone option at configure time.
Defaults to disabled when --enable-tcg-plugin is not set.
When --enable-tcg-plugin is set and the capstone library
is not found or of incompatible version, the dyncount
plugin is a no-op and it emits aa warning at runtime.