tcg/plugins/dyncount: fix management of capstone instructions ids
Actual instruction ids may be greater than the declared max
static instruction id for some architecture implementations.

In particular on aarch64, some internal ids are used for instruction
aliases. In this case we add the instruction as an _unknown_
mnemonic.

Note that capstone API cs_insn_name() is bogus in this case as it
returns a null string.
For instance on aarch64, insn->id == 65535 is for
the negs mnemonic, but cs_insn_name() return the null string.